# openapi/roomSync.yaml
openapi: 3.0.3
info:
  title: RoomSync UT API
  version: 1.0.0
servers:
  - url: https://api.roomsync.local
  - url: http://localhost:4000
paths:
...
  /listings:
    get:
      summary: List marketplace listings
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [sublet, sublets, furniture, textbooks, parking, other] }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: min
          schema: { type: number }
        - in: query
          name: max
          schema: { type: number }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [price_asc, price_desc, date_asc, date_desc] }
      responses: { '200': { description: OK } }
    post:
      summary: Create listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, title, price]
              properties:
                type: { type: string, enum: [sublet, sublets, furniture, textbooks, parking, other] }
                categories: { type: array, items: { type: string } }
                title: { type: string }
                description: { type: string }
                price: { type: number }
                photos: { type: array, items: { type: string } }
                photosBase64: { type: array, items: { type: string } }
                loc: { type: object, properties: { lat: { type: number }, lng: { type: number } } }
                availableFrom: { type: string, format: date-time }
                availableTo: { type: string, format: date-time }
                status: { type: string, enum: [available, pending, sold] }
      responses: { '200': { description: OK } }
  /listings/{id}:
    patch:
      summary: Update listing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses: { '200': { description: OK } }
  /listings/{id}/favorite:
    post:
      summary: Favorite listing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: OK } }
  /listings/{id}/unfavorite:
    post:
      summary: Unfavorite listing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: OK } }
...